gridtime: move to 1d datasets as well
