+++ trajan/traj.py
@@ -302,3 +302,64 @@ class Traj:
             area.append(hull.volume)  # volume=area for 2D as here
 
         return np.array(area)
+
+    def gridtime(self, times):
+        """Interpolate dataset to regular time interval
+
+        times:
+            - an array of times, or
+            - a string "freq" specifying a Pandas daterange (e.g. 'h', '6h, 'D'...)
+
+        Note that the resulting DataSet will have "time" as a dimension coordinate.
+        """
+
+        from scipy.interpolate import interp1d
+        import pandas as pd
+
+        if isinstance(times, str):  # Make time series with given interval
+            freq = times
+            start_time = np.nanmin(np.asarray(self.ds.time))
+            start_time = pd.to_datetime(start_time).strftime('%Y-%m-%d')
+            end_time = np.nanmax(np.asarray(self.ds.time)) + \
+                            np.timedelta64(23, 'h') + np.timedelta64(59, 'm')
+            end_time = pd.to_datetime(end_time).strftime('%Y-%m-%d')
+            times = pd.date_range(start_time, end_time, freq=freq)
+
+        # Why not..
+        if not isinstance(times, np.ndarray):
+            times = times.to_numpy()
+
+        # Create empty dataset to hold interpolated values
+        d = xr.Dataset(coords={
+            'trajectory': self.ds['trajectory'],
+            'time': (["time"], times)
+        },
+                       attrs=self.ds.attrs)
+
+        for varname, var in self.ds.variables.items():
+            if varname in ('time', 'obs'):
+                continue
+            if 'obs' not in var.dims:
+                d[varname] = var
+                continue
+
+            # Create empty dataarray to hold interpolated values for given variable
+            da = xr.DataArray(data=np.zeros(
+                tuple(d.dims[di] for di in ['trajectory', 'time'])) * np.nan,
+                              dims=d.dims,
+                              coords=d.coords,
+                              attrs=var.attrs)
+
+            origtimes = self.ds['time'].ffill(dim='obs').astype(np.float64)
+
+            for t in range(
+                    self.ds.dims['trajectory']):  # loop over trajectories
+                # Make interpolator
+                f = interp1d(origtimes.isel(trajectory=t), var.isel(trajectory=t), bounds_error=False)
+                # Interpolate onto given times
+                da.loc[{'trajectory': t}] = f(times.astype(np.float64))
+
+            d[varname] = da
+
+        return d
+
